Refactor solr tests to submodules #21
